Improve handling of GraphQL errors.
Always throw an error, even if there's partial data.  Attach any data to
the error, so the onError handler can use it (and return it) if desired.

Map some known error types to statuses (403, 404), and construct messages
that should be compatible with REST message-sniffing code.

Return data directly as the result, since if there are any errors they're
now handled entirely as an exception.

Use logTag to provide context on error origin and improve fingerpriting.