Add a GraphQL preprocessor for adjusting queries on the fly.
While GraphQL queries can be parametrized in various ways (variables,
directives such as `@include`, type switches, etc.), the spec requires
that all queries be completely valid -- even parts that won't be
executed.  This means that you if you reference fields not defined in
the schema the query will fail validation, which is a problem if you're
writing queries that must work with multiple schema versions.

The preprocessor works around this limitation by introducing a syntax
to conditionally include/exclude parts of the query based on either the
GHE version or the user's currently authorized scopes.  (The latter is
not strictly necessary but can be convenient when dealing with type
switches.)

The preprocessor syntax is based on GraphQL comments.  This runs the
risk of the preprocessor directives being accidentally ignored, but
retains compatibility with GraphQL syntax checkers and allows the
queries to be pasted into interactive explorers as-is.