Add `stale` flag and `scope()` method.
The `stale` flag allows for reuse of expired cached data, in cases where
we don't care about freshness at all.  It's stronger than `fresh: false`
in that respect.

The `scope()` method is just a convenient way of augmenting the options
of an existing Hubkit object, paralleling the API in NodeFire.